FIO-8343: added comments, moved tests to separate file
